Fix to initial template of LSP vscode extension, will do all the server code inside it.
Still figuring out what I need to make it work.

I will need the LSP in order to make the virtual view work when someone clicks on a tag / class / id to open it, showing all the concatenated CSS in the view.

Make use of the workspace to get all the CSS files and ahead of time make all the virtual views there could be, or cache them when selected.
Will need to see about highlighting background to indicate what file something is coming from. GitHub extension does something similar for highlighting changes.

Not for Project Log;
Struggling with motivation to even get the work done, am still on track to get a first (somehow?). I probably will have nothing to show for my meeting tomorrow.