Getting the client and server working for Proof of concept
Currently the server is crashing at the moment, need to figure that out. Proof of concept will consist of the server being able to concatenate CSS files in an html server to a new window. If I can, I'll try get it to save back to the original files if changes are made.