Adding in some global state
Added Global state to the server,
this was done so that we don't need to check where the workspace paths are every time we open a file.

there are checks that need to be done to update the workspace paths when it changes, but otherwise, it doesn't change and only needs to be fetched at the start of the server initialisation.

Need to work on the saving feature as well. It would be worth making a list of features that needs to still be done so I can cross them off.