Lots of Changes - Check Description
Cargo.toml -
Added chrono, for time purposes
Added once_cell, for Lazy initialisation of the global data
Added path-absolutize, for being able to get the absolute path given another absolute path and a relative path
Added tokio-test, able to unit test async methods

extension.ts - Added an activate function just so it's easier to attach the debugger to the server when it starts.

Added rustfmt.toml - Help stylise the document

File.rs - removed all async from it, limiting async to the main.rs, removed the Files trait as it was unecessary

main.rs - There's been a lot of fighting with trying to add some mutable static data to the server. The mutable data makes it easier in terms of getting workspaces as they rarely change, and when they do, it is tracked anyway.
Basically a problem is that because a Mutex blocks access to the memory, and asynchronous calls could theoretically take forever, it means we cannot guarantee when a mutex will be unlocked by the time we need to access it again, there was a race condition being formed because the initialized housed the call to get the workspaces at start up but we also started the LSP by opening a file, so it was opening the file simultaneously to the initialized status, meaning on the first run, the workspaces would have not been found so when we try to open a file, we can't find where it is because we don't know what workspace it is in because it hasn't been committed to memory yet. the solution was to do a check if the memory has been initialized every time a file gets opened, and this prevents the race condition.
It is inside an Arc (Atomic Reference Count), which in all honesty, probably doesn't do anything, but it's in a working state right now so I don't want to touch it anymore.

Memory added an is_ready boolean to its struct for the first time check.

setting up the metadata file. Metadata will be used to be able to remember what file a style came from and saving back, what file it should go to. Currently the serializing and deserializing works, but need to implement its logic into the app.