Creating the metadata files.
Currently just getting the groundwork in place for the metadata

The metadata will be used to keep track of styles and which they belong to.
The idea being that each html file SHOULD ideally have only one link href to a css file, and we use the power of this extension to keep track of importing shared css styles into the individual sheets.

I'll need to have a fallback for when a html file has more than one, probably dump the file into the .bhc folder