Lots of changes
Extension.ts + package.json
Now only starts if there is .html or .css in the workspace, rather than when a file is opened.

file.rs -> removing .to_string() where it was unncessary
main.rs -> Got the css file to json partially working, still need to update the meta.json, and get the html json files being made
memory.rs -> Removing adding temp_dir() to the list of workspaces if there are no workspaces currently open. The extension will only run when a workspace is open, so we don't need this edge case anymore.
metadata.rs -> split json_files into a struct of Vecs containing specific json files for html, css, and general json, this is to avoid overriding other json files that aren't to do with the extension.

css_metadata.rs -> Migrated a lot of the code into separate files, i.e. the nested structs, files, styles, and attributes. updated some unit tests to actual cover code.

css_attribute.rs -> Moved from css_metadata.rs, added a function to update or insert a value.
value has been changed to values (String -> Vec<String>), this is because an attribute could have multiple declarations, and we don't want to remove one at random despite it being wrong as the last to be declared will be the one to apply its style. The goal of this LSP is not to provide corrections for wrong code (as of right now)

css_style.rs -> moved from css_metadata.rs
update_attributes, this function takes in a list of new attributes, the new attributes are from the stylesheet in the state it is curently in, the old is the metadata before being updated, we need to update the attributes to reflect the stylesheet.
an attribute can be:
* Added, therefore we need to add a new attribute to the style
* Modified, therefore we need to change an existing attribute in the style
* Removed, therefore we need to remove the attribute from the style.
This needs to go into such detail because the css file does not remember where a style has come from if it has been imported from another sheet, so we need to iteratively go through all of them to not lose that information, later down the line, this will need to check when an attribute has been changed, and was taken from another source, and if it has been changed back... therefore, no longer overwritten, but still from another source.

workspace_metadata.rs -> made helper functions to get the next available id for a css / html file, a file might be deleted, therefore the metadata file must be deleted too, which will knock the order off, it doesn't make sense to move all the existing files, so instead these functions find when there is a missing number in the order when creating a new file and gives it to the new metadata file.

workspace_css_file.rs -> moved from workspace_metadata
workspace_html_file.rs -> moved from workspace_metadata